requirements.txt
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.14.0
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dateutil>=2.8.0

